requirements.txt
# ─────────────────────────────────────────────
# MedSafe AI – Python Dependencies
# ─────────────────────────────────────────────

# Web framework & Dashboard
Flask==3.1.0
streamlit==1.41.1

# Kafka streaming
kafka-python==2.0.2

# PostgreSQL
psycopg2-binary==2.9.10

# Data science & ML
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
tensorflow==2.18.0

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# Fuzzy string matching
rapidfuzz==3.11.0

# Plotting
matplotlib==3.10.0
plotly==5.24.1

# Utilities
requests==2.32.3
python-dateutil==2.9.0
